requirements.txt
# Basic dependencies
python-dotenv
anthropic==0.45.0



pandas==2.2.1
matplotlib==3.8.3
ipython>=8.12.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# ML dependencies
torch>=2.1.0
transformers==4.48.1
peft==0.7.1
bitsandbytes==0.43.2
accelerate==0.26.0
datasets==3.2.0

# CUDA dependencies
nvidia-cuda-runtime-cu11>=11.8
nvidia-cudnn-cu11>=8.9.0

# Development dependencies
flake8>=6.1.0 